+++ ostap/core/core.py
@@ -58,6 +58,8 @@ __all__     = (
     ##
     'loop_items'       , ## loop over dictionary items 
     'items_loop'       , ## ditto
+    ##
+    'is_sorted'        , ## check that list is sorted 
     )
 # =============================================================================
 import math, sys, os 
@@ -455,6 +457,15 @@ def _rd_valid_ ( rdir ) :
         
 ROOT.TDirectory.__bool__     = _rd_valid_
 ROOT.TDirectory.__nonzero__  = _rd_valid_
+
+
+# =============================================================================
+## check that list is sorted 
+def is_sorted ( lst ) :
+    """Check that list is sorted
+    """
+    l = len ( lst )  
+    return all ( lst[i] <= lst[i+1] for i in range ( l - 1 ) ) if lst else True 
 
 # =============================================================================
 ## define the build directory for ROOT