+++ ostap/parallel/parallel_statvar.py
@@ -70,44 +70,69 @@ class StatVarTask(Task) :
         - number of entries to process
         """
 
+        print ('I AM HERE START/1', jobid, item ) 
         import ROOT
         from ostap.logger.utils import logWarning
         with logWarning() :
             import ostap.core.pyrouts 
             import ostap.trees.trees 
 
+        print ('I AM HERE START/2', jobid, item )
+        
         chain   = item.chain 
         first   = item.first
-        ## last    = min ( n_large , first + item.nevents if 0 < item.nevents else n_large )
-        last    = n_large
+        last    = min ( n_large , first + item.nevents if 0 < item.nevents else n_large )
         
+        print ('I AM HERE START/3', jobid, item )
+
         from ostap.trees.trees  import _stat_vars_
+
+        print ('I AM HERE START/4', jobid , self.what , self.cuts , first , last , type ( chain )  )
+
         self.__output = _stat_vars_ ( chain , self.what , self.cuts , first , last )
 
+        print ('I AM HERE END', jobid, item )
+        
         return self.__output 
         
     ## merge results 
     def merge_results ( self , result , jobid = -1 ) :
         
+        print ('I AM HERE MERGE/1', jobid )
+
         from ostap.stats.counters import WSE
 
+        
         if not self.__output : self.__output = result
         else               :
+            print ('I AM HERE MERGE/2', jobid )
+            
             assert type( self.__output ) == type ( result ) , 'Invalid types for merging!'
-            if isinstance ( self.__output , dict ) : 
+
+            print ('I AM HERE MERGE/3', jobid )
+
+            if isinstance ( self.__output , dict ) :
+                print ('I AM HERE MERGE/4', jobid , result.keys() )
+
                 for key in result : 
-                    if self.output.has_key ( key ) : self.__output[key] += result[key]
-                    else                           : self.__output[key]  = result[key] 
+                    if    key in self.output : self.__output [ key ] += result [ key ]
+                    else                     : self.__output [ key ]  = result [ key ] 
             else :
+                
+                print ('I AM HERE MERGE/5', jobid )
+                
                 self.__output += result
+
+            print ('I AM HERE MERGE/6', jobid )
 
     ## get the results 
     def results ( self ) : return self.__output 
 
+
 # ===================================================================================
 ## parallel processing of loooong chain/tree 
 #  @code
-#  chain    = ...
+#  chain          = ...
 #  chain.pStatVar ( .... ) 
 #  @endcode 
 def pStatVar ( chain               ,
@@ -115,27 +140,33 @@ def pStatVar ( chain               ,
                cuts       = ''     ,
                nevents    = -1     ,
                first      =  0     ,
-               chunk_size = 100000 ,
-               max_files  = 10     ,
+               chunk_size = 250000 ,
+               max_files  =  1     ,
                silent     = True   , **kwargs ) :
     """ Parallel processing of loooong chain/tree 
     >>> chain    = ...
     >>> chain.pstatVar( 'mass' , 'pt>1') 
     """
 
+    print ( 'I am pStatVar/1' ) 
     ## few special/trivial cases
 
     last = min ( n_large , first + nevents if 0 < nevents else n_large )
     
-
     if 0 <= first and 0 < nevents < chunk_size :
+        print ( 'I am pStatVar/1.1' ) 
         return chain.statVar ( what , cuts , first , last )
-    elif isinstance ( chain , ROOT.TChain ) : 
-        if chain.nFiles() < 5 and len ( chain ) < chunk_size :
+    elif isinstance ( chain , ROOT.TChain ) :
+        print ( 'I am pStatVar/1.2' ) 
+        if chain.nFiles() < 3 and len ( chain ) < chunk_size :
+            print ( 'I am pStatVar/1.3' ) 
             return chain.statVar ( what , cuts , first , last )                         
     elif isinstance ( chain , ROOT.TTree  ) and len ( chain ) < chunk_size :
+        print ( 'I am pStatVar/1.4' ) 
         return chain.statVar ( what , cuts , first , last ) 
-    
+
+    print ( 'I am pStatVar/2' ) 
+
     from ostap.trees.trees import Chain
     ch     = Chain ( chain , first = first , nevents = nevents )
 
@@ -144,7 +175,11 @@ def pStatVar ( chain               ,
 
     trees  = ch.split ( chunk_size = chunk_size , max_files = max_files )
 
+    print ( 'I am pStatVar/3' , len(trees)  ) 
+
     wmgr.process ( task , trees )
+
+    print ( 'I am pStatVar/4' , len(trees)  ) 
 
     del trees
     del ch    
@@ -175,7 +210,7 @@ if '__main__' == __name__ :
     
     if not ( 2**32 - 1 ) <= n_large <= ROOT.TVirtualTreePlayer.kMaxEntries :
         logger.error ( "Invalid setting of ``n_large''(%d) parameter (>%d)" % ( n_large , ROOT.TVirtualTreePlayer.kMaxEntries ) )
-        
+
 # =============================================================================
 #                                                                       The END 
 # =============================================================================