+++ source/CMake_OSTAP.cmake
@@ -17,6 +17,7 @@ add_library(ostap SHARED src/format.cpp
                          src/Chi2Fit.cpp
                          src/Dalitz.cpp
                          src/DalitzIntegrator.cpp
+                         src/DataFrameActions.cpp
                          src/DataFrameUtils.cpp
                          src/EigenSystem.cpp   
                          src/Error2Exception.cpp